+++ packages/linkfive_purchases/lib/client/linkfive_billing_client_interface.dart
@@ -4,6 +4,6 @@ import 'package:linkfive_purchases/models/linkfive_response.dart';
 /// Internal Billing Client. It holds the connection to the native billing sdk
 abstract class LinkFiveBillingClientInterface {
   /// load the products from the native billing sdk
-  Future<List<ProductDetails>?> getPlatformSubscriptions(
+  Future<List<ProductDetails>?> getPlatformProducts(
       LinkFiveResponseData linkFiveResponse);
 }