+++ packages/linkfive_purchases/lib/client/linkfive_client_interface.dart
@@ -1,8 +1,6 @@
 import 'package:in_app_purchase_android/in_app_purchase_android.dart';
 import 'package:in_app_purchase_storekit/in_app_purchase_storekit.dart';
 import 'package:linkfive_purchases/linkfive_purchases.dart';
-import 'package:linkfive_purchases/models/linkfive_restore_apple_item.dart';
-import 'package:linkfive_purchases/models/linkfive_restore_google_item.dart';
 import 'package:linkfive_purchases/store/linkfive_app_data_store.dart';
 
 /// HTTP client to LinkFive
@@ -14,26 +12,32 @@ abstract class LinkFiveClientInterface {
 
   /// after a purchase on ios we call the purchases/apple
   /// We don't need to do this on Android
-  Future<List<LinkFivePlan>> purchaseIos(AppStoreProductDetails productDetails,
-      AppStorePurchaseDetails purchaseDetails);
+  Future<LinkFiveActiveProducts> purchaseIos({
+    required AppStoreProductDetails? productDetails,
+    required AppStorePurchaseDetails purchaseDetails,
+  });
 
   /// after a purchase on Google we call the purchases/google
   /// We don't need to do this on Android
-  Future<List<LinkFivePlan>> purchaseGooglePlay(
+  Future<LinkFiveActiveProducts> purchaseGooglePlay(
       GooglePlayPurchaseDetails purchaseDetails, GooglePlayProductDetails productDetails);
+
+  /// after a one time purchase purchase on Google we post the data to LinkFIve
+  Future<LinkFiveActiveProducts> purchaseGooglePlayOneTimePurchase(
+      GooglePlayPurchaseDetails purchaseDetails, OneTimePurchaseOfferDetailsWrapper otpDetails);
 
   /// Fetches the receipts for a user
   ///
   /// if no LinkFive UUID is provided, LinkFive will generate a new user ID
   ///
-  Future<List<LinkFivePlan>> fetchUserPlanListFromLinkFive();
+  Future<LinkFiveActiveProducts> fetchUserPlanListFromLinkFive();
 
   /// RESTORE APPLE APP STORE
   ///
   /// This will send all restored transactionIds to LinkFive
   /// We will check against apple if those transaction are valid and
   /// enable or disable a product
-  Future<List<LinkFivePlan>> restoreIos(
+  Future<LinkFiveActiveProducts> restoreIos(
       List<LinkFiveRestoreAppleItem> restoreList);
 
   /// RESTORE GOOGLE PLAY STORE
@@ -41,9 +45,9 @@ abstract class LinkFiveClientInterface {
   /// This will send all restored transactionIds to LinkFive
   /// We will check against apple if those transaction are valid and
   /// enable or disable a product
-  Future<List<LinkFivePlan>> restoreGoogle(
+  Future<LinkFiveActiveProducts> restoreGoogle(
       List<LinkFiveRestoreGoogleItem> restoreList);
 
   /// Should change the USER ID.
-  Future<List<LinkFivePlan>> changeUserId(String? userId);
+  Future<LinkFiveActiveProducts> changeUserId(String? userId);
 }