+++ packages/linkfive_purchases/lib/logic/linkfive_purchases_impl.dart
@@ -18,8 +18,6 @@ import 'package:linkfive_purchases/logic/extensions/initialize_extension.dart';
 import 'package:linkfive_purchases/linkfive_purchases.dart';
 import 'package:linkfive_purchases/logic/linkfive_user_management.dart';
 import 'package:linkfive_purchases/logic/upgrade_downgrade_purchases.dart';
-import 'package:linkfive_purchases/models/linkfive_restore_apple_item.dart';
-import 'package:linkfive_purchases/models/linkfive_restore_google_item.dart';
 import 'package:linkfive_purchases/store/linkfive_app_data_store.dart';
 import 'package:linkfive_purchases/store/linkfive_prefs.dart';
 import 'package:linkfive_purchases/store/linkfive_store.dart';
@@ -65,6 +63,9 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
 
   /// Cache of AppStoreProductDetail
   ProductDetails? _productDetailsToPurchase;
+
+  /// the v2 of purchases waits for the async purchase response and completes the future
+  final List<Completer<LinkFiveActiveProducts>> registeredPurchaseCompleterList = [];
 
   /// LinkFive HTTP Client
   ///
@@ -130,7 +131,7 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
       LinkFiveEnvironment env = LinkFiveEnvironment.PRODUCTION}) async {
     if (appDataStore.apiKey.isNotEmpty) {
       LinkFiveLogger.w("LinkFive is already Initialized. Please only call init once");
-      return _store.latestLinkFiveActiveProducts ?? LinkFiveActiveProducts();
+      return _store.latestLinkFiveActiveProducts ?? LinkFiveActiveProducts.empty();
     }
 
     LinkFiveLogger.setLogLevel(logLevel);
@@ -193,9 +194,9 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
     var linkFiveResponse = await _client.fetchLinkFiveResponse();
     _store.onNewResponseData(linkFiveResponse);
 
-    List<ProductDetails>? platformSubscriptions = await _billingClient.getPlatformSubscriptions(linkFiveResponse);
-    if (platformSubscriptions != null) {
-      return _store.onNewPlatformSubscriptions(platformSubscriptions);
+    List<ProductDetails>? platformProducts = await _billingClient.getPlatformProducts(linkFiveResponse);
+    if (platformProducts != null) {
+      return _store.onNewPlatformProduct(platformProducts);
     }
     return null;
   }
@@ -283,6 +284,19 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
     return showBuySuccess;
   }
 
+  /// USE WITH CARE. We're currently testing the new purchase method
+  ///
+  /// All In App purchases are handled async, it's difficult to work with futures.
+  ///
+  /// We're waiting for the next purchase update and complete the future.
+  Future<LinkFiveActiveProducts> purchaseFuture(dynamic productDetails) async {
+    makeSureIsInitialize();
+    final completer = Completer<LinkFiveActiveProducts>();
+    registeredPurchaseCompleterList.add(completer);
+    await purchase(productDetails);
+    return completer.future;
+  }
+
   Future<LinkFiveActiveProducts> reloadActivePlans() async {
     await waitForInit();
     return await _updateActivePlansFromLinkFive();
@@ -297,6 +311,19 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
     super.isPendingPurchase = true;
     await inAppPurchaseInstance.restorePurchases();
     return false;
+  }
+
+  /// USE WITH CARE. We're currently testing the new restore method
+  ///
+  /// All In App purchases are handled async, it's difficult to work with futures.
+  ///
+  /// We're waiting for the next purchase update and complete the future.
+  Future<LinkFiveActiveProducts> restoreWithFuture() async {
+    makeSureIsInitialize();
+    final completer = Completer<LinkFiveActiveProducts>();
+    registeredPurchaseCompleterList.add(completer);
+    await restore();
+    return completer.future;
   }
 
   /// Handles the Up and Downgrade of a Subscription plans
@@ -346,17 +373,17 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
     // do the LinkFive request
     if (userId != previousUserId) {
       LinkFiveLogger.d("User Id has changed since last time update LinkFive");
-      final List<LinkFivePlan> linkFivePlanList = await _client.changeUserId(userId);
+      final LinkFiveActiveProducts activeProducts = await _client.changeUserId(userId);
 
       // notify all listeners
-      final activeProducts = _store.onNewLinkFiveNewActivePlanList(linkFivePlanList);
+      _store.onNewLinkFiveNewActiveProducts(activeProducts);
 
       // update purchase State
-      super.updateStateFromActivePlanList(linkFivePlanList);
+      super.updateStateFromActiveProducts(activeProducts);
 
       return activeProducts;
     }
-    return _store.latestLinkFiveActiveProducts ?? LinkFiveActiveProducts();
+    return _store.latestLinkFiveActiveProducts ?? LinkFiveActiveProducts.empty();
   }
 
   //#region private methods
@@ -369,13 +396,13 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
     LinkFiveLogger.d("Update active plans from LinkFive");
     try {
       // Fetch all Plans from LinkFive
-      final List<LinkFivePlan> linkFivePlanList = await _client.fetchUserPlanListFromLinkFive();
+      final activeProducts = await _client.fetchUserPlanListFromLinkFive();
 
       // notify all listeners
-      final activeProducts = _store.onNewLinkFiveNewActivePlanList(linkFivePlanList);
+      _store.onNewLinkFiveNewActiveProducts(activeProducts);
 
       // update purchase State
-      super.updateStateFromActivePlanList(linkFivePlanList);
+      super.updateStateFromActiveProducts(activeProducts);
 
       // return with the created object
       return activeProducts;
@@ -384,7 +411,7 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
     }
 
     // return an empty List on error
-    return LinkFiveActiveProducts();
+    return LinkFiveActiveProducts.empty();
   }
 
   ///
@@ -442,15 +469,19 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
           _productDetailsToPurchase = null;
 
           // handle ios Purchase
-          if (Platform.isIOS && productDetails != null) {
-            final appstorePurchaseDetails = purchaseDetails as AppStorePurchaseDetails;
-
+          if (Platform.isIOS) {
             // handle the ios purchase request to LinkFive
-            await _handlePurchaseApple(appstorePurchaseDetails, productDetails as AppStoreProductDetails);
+            await _handlePurchaseApple(
+              appstorePurchaseDetails: purchaseDetails as AppStorePurchaseDetails,
+              productDetailsToPurchase: productDetails,
+            );
           }
           // Handle google play purchase
           if (purchaseDetails is GooglePlayPurchaseDetails) {
-            await _handlePurchaseGoogle(purchaseDetails, productDetails as GooglePlayProductDetails);
+            await _handlePurchaseGoogle(
+              purchaseDetails: purchaseDetails,
+              productDetails: productDetails as GooglePlayProductDetails,
+            );
           }
 
           super.isPendingPurchase = false;
@@ -475,6 +506,12 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
       await _handleRestore(purchaseDetailsList);
       super.isPendingPurchase = false;
     }
+
+    // complete the purchase future
+    for (final completer in registeredPurchaseCompleterList) {
+      completer.complete(_store.latestLinkFiveActiveProducts);
+    }
+    registeredPurchaseCompleterList.clear();
   }
 
   /// This will handle the restore functionality
@@ -482,26 +519,28 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
   /// we will send all transactions to LinkFive and connect them to the
   /// current user
   Future<void> _handleRestore(List<PurchaseDetails> purchaseDetailsList) async {
-    List<LinkFivePlan> restoredPlans = [];
+    LinkFiveActiveProducts? activeProducts;
 
     // handle each platform separately
     if (Platform.isIOS) {
-      restoredPlans = await _handleRestoreApple(purchaseDetailsList);
+      activeProducts = await _handleRestoreApple(purchaseDetailsList);
     } else if (Platform.isAndroid) {
-      restoredPlans = await _handleRestoreGoogle(purchaseDetailsList);
+      activeProducts = await _handleRestoreGoogle(purchaseDetailsList);
     }
 
+    activeProducts ??= LinkFiveActiveProducts.empty();
+
     // update purchase state also from the restore function
-    super.updateStateFromActivePlanList(restoredPlans);
+    super.updateStateFromActiveProducts(activeProducts);
 
     // notify all listeners
-    _store.onNewLinkFiveNewActivePlanList(restoredPlans);
+    _store.onNewLinkFiveNewActiveProducts(activeProducts);
   }
 
   ///
   /// Restore Apple
   ///
-  Future<List<LinkFivePlan>> _handleRestoreApple(List<PurchaseDetails> purchaseDetailsList) async {
+  Future<LinkFiveActiveProducts> _handleRestoreApple(List<PurchaseDetails> purchaseDetailsList) async {
     final List<LinkFiveRestoreAppleItem> restoredTransactionList = [];
 
     // check each item if it is not null and save the transactionId
@@ -523,7 +562,7 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
     if (restoredTransactionList.isNotEmpty) {
       return await _client.restoreIos(restoredTransactionList);
     }
-    return [];
+    return LinkFiveActiveProducts.empty();
   }
 
   /// This will handle the Google restore functionality
@@ -533,7 +572,7 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
   ///
   /// @return empty list if there is no subscription
   ///
-  Future<List<LinkFivePlan>> _handleRestoreGoogle(List<PurchaseDetails> purchaseDetailsList) async {
+  Future<LinkFiveActiveProducts> _handleRestoreGoogle(List<PurchaseDetails> purchaseDetailsList) async {
     final List<LinkFiveRestoreGoogleItem> restoredList = [];
 
     // check each item if it is not null and save the purchaseID
@@ -551,34 +590,54 @@ class LinkFivePurchasesImpl extends DefaultPurchaseHandler implements CallbackIn
     if (restoredList.isNotEmpty) {
       return await _client.restoreGoogle(restoredList);
     }
-    return [];
+    return LinkFiveActiveProducts.empty();
   }
 
   ///
   /// This will handle the apple purchase process and update all listeners
   ///
-  _handlePurchaseApple(
-      AppStorePurchaseDetails appstorePurchaseDetails, AppStoreProductDetails productDetailsToPurchase) async {
-    List<LinkFivePlan> linkFivePlanList = await _client.purchaseIos(productDetailsToPurchase, appstorePurchaseDetails);
+  _handlePurchaseApple({
+    required AppStorePurchaseDetails appstorePurchaseDetails,
+    required ProductDetails? productDetailsToPurchase,
+  }) async {
+    final linkFiveActiveProducts = await _client.purchaseIos(
+      productDetails: switch (productDetailsToPurchase) {
+        null => null,
+        _ => productDetailsToPurchase as AppStoreProductDetails
+      },
+      purchaseDetails: appstorePurchaseDetails,
+    );
 
     // update purchase State
-    super.updateStateFromActivePlanList(linkFivePlanList);
+    super.updateStateFromActiveProducts(linkFiveActiveProducts);
 
     // notify all listeners
-    _store.onNewLinkFiveNewActivePlanList(linkFivePlanList);
+    _store.onNewLinkFiveNewActiveProducts(linkFiveActiveProducts);
   }
 
   ///
   /// This will handle the Google Play purchase process and update all listeners
   ///
-  _handlePurchaseGoogle(GooglePlayPurchaseDetails purchaseDetails, GooglePlayProductDetails productDetails) async {
-    List<LinkFivePlan> linkFivePlanList = await _client.purchaseGooglePlay(purchaseDetails, productDetails);
+  _handlePurchaseGoogle({
+    required GooglePlayPurchaseDetails purchaseDetails,
+    required GooglePlayProductDetails productDetails,
+  }) async {
+    if (productDetails.productDetails.oneTimePurchaseOfferDetails != null) {
+      final activeProducts = await _client.purchaseGooglePlayOneTimePurchase(
+          purchaseDetails, productDetails.productDetails.oneTimePurchaseOfferDetails!);
 
-    // update purchase State
-    super.updateStateFromActivePlanList(linkFivePlanList);
+      super.updateStateFromActiveProducts(activeProducts);
 
-    // notify all listeners
-    _store.onNewLinkFiveNewActivePlanList(linkFivePlanList);
+      _store.onNewLinkFiveNewActiveProducts(activeProducts);
+    } else if (productDetails.productDetails.subscriptionOfferDetails != null) {
+      final activeProducts = await _client.purchaseGooglePlay(purchaseDetails, productDetails);
+
+      // update purchase State
+      super.updateStateFromActiveProducts(activeProducts);
+
+      // notify all listeners
+      _store.onNewLinkFiveNewActiveProducts(activeProducts);
+    }
   }
 
   /// This method overrides the interface to fetch the subscriptions on view.