+++ packages/linkfive_purchases/lib/models/product_type.dart
@@ -1,0 +1,3 @@
+enum LinkFiveProductType {
+  Subscription, OneTimePurchase;
+}